implémentation de la classe Volonté

+++ src/main/java/testament/dao/UserRepository.java
@@ -3,6 +3,6 @@ package testament.dao;
 import testament.entity.Utilisateur;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<Utilisateur, Long> {
+public interface UserRepository extends JpaRepository<Utilisateur, Integer> {
     Utilisateur findByUsername(String name);
 }
